% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PU_EMLasso.R
\name{PU_EMLasso}
\alias{PU_EMLasso}
\title{PU learning solved by regularized (LASSO) EM algorithm}
\usage{
PU_EMLasso(
  Z,
  X,
  PI,
  lambda.sel = "lambda.min",
  intercept = TRUE,
  eps = 0.001,
  max.iter = 1000,
  ...
)
}
\arguments{
\item{Z}{The observed binary outcomes under the PU setting, i.e. the 1s represent the positives instances with certainty and the 0s represent the unlabeled instances}

\item{X}{The nxp design matrix, where n is the number of instances and the p is the number of covariates (no bias/intercept term, i.e., the whole 1 column)}

\item{PI}{The population prevalence, i.e., Pr(y==1), where y is the true binary outcomes}

\item{lambda.sel}{c('lambda.min', 'lambda.1se'), used for selecting lambda in cv.glmnet. The default is 'lambda.min'}

\item{intercept}{If intercept should be considered. The default is TRUE}

\item{eps}{Convergence threshold for EM algorithm}

\item{max.iter}{Maximum number of iterations for the EM algorithm}

\item{...}{Other arguments for glmnet/cv.glmnet}
}
\value{
A list object containing estimated probability, estimated/selected parameters, and the corresponding penalty coefficient lambda (based on glmnet).
}
\description{
General parametric framework for the PU learning with variable selection embedded
}
\details{
This is the PU learning method proposed in the paper: Ward, G., Hastie, T., Barry, S., Elith, J., & Leathwick, J. R. (2009). Presence‐only data and the EM algorithm. Biometrics, 65(2), 554-563
}
\examples{
set.seed(123)
res = PU_EMLasso(sample_dat$Z, sample_dat$X, sample_dat$PI, alpha = 1)
# plot(res$Est.Pr, sample_dat$true.prob)
}
