% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dat.R
\docType{data}
\name{sample_dat}
\alias{sample_dat}
\title{Sample toy data to illustrate the package usage}
\format{
A list of 4 items. Z: observed binary outcomes, either 0 or 1. X: covariates; PI: population prevalence; true.prob: the true probablity of being positive (1) for each instance.
}
\description{
Sample toy data to illustrate the package usage
}
